feat: 게시물 필터 영역을 PostFilterSection 컴포넌트로 분리
